requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
tqdm>=4.64.0

# Deep Learning and NLP
torch>=1.13.0
transformers>=4.21.0
sentence-transformers>=2.2.0

# Nouveaux pour le moteur conceptuel v2.4
gensim>=4.2.0
spacy>=3.4.0

# LlamaIndex ecosystem
llama-index>=0.9.0
llama-index-llms-openai>=0.1.0

# API and web
openai>=1.44.1
httpx>=0.24.0

# Text processing
tiktoken>=0.3.0

# Data handling
pathlib2>=2.3.0
pickle5>=0.0.11; python_version < "3.8"

# Optional: for better performance
faiss-cpu>=1.7.0  # Alternative to scikit-learn for vector search

# Development dependencies (optional)
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
pathlib
